Fix gradle issue.
For internal builds we don't want to reference mavenCentral, so this
CL splits the build.gradle in 2 files. The main one only sets up
the classpath of the build using mavenCentral, while the other contains
the actual build rules (rules.gradle).

The main file also applies the rules file so that no manual setup
is required. It's all transparent for a standalone project.

When building internally, we'll redirect the build to use rules.gradle
directly instead of build.gradle, bypassing the mavenCentral setup.

Change-Id: I7d334a5ddbdf40a856bf24ca9f41d3b8f61230ff