Add a getImageURL method
Adds a method that allows a user to get the URL of the image set in the NetworkImageView.
This, for example, may be useful to check if a NetworkImageView has an image defined to be loaded with only access to the view itself. It may be necessary when the user doesn't have access to the original URL in their scope.