Add the option of setting DiskBasedCache sizes
The default DiskBasedCache size of 5MB is not appropriate for all users of volley.
There is no interface for users to set this size when using the standard volley implementation. Before this change, one would need to fork their own Volley.java file to be able to change the size of the DiskBasedCache.