Add JaCoCo for code coverage reports
It looks that emma reports are not perfect.
Some classes are missed in the report.

JaCoCo does the job. And you can define rules
to break the build if the coverage to low.
Currently the limit is 40% for the whole project.
This is what Volley currently has.

To get the report call (one of them):
$ mvn clean verify
$ mvn clean install

You find the report in target\jacoco-report

Change-Id: I7f5e2d5de0ffaa779148dfb0e239e8e95a1055e0
Signed-off-by: Ralph Bergmann <ralph@the4thfloor.eu>